feat(gpu): WGPU X11 宿主切片——WgpuX11Surface 组合宿主接线、WSLg 真机验收与三处口径修正
- 新增 WgpuX11Surface（组合内嵌 X11Surface，零 Xlib 复制）：GPU 帧路径 +
  XPutImage 软件回退分流，失败分层与帧契约同 Win32 WgpuSurface
- 工厂路由：WgpuOptions/create_window 门控改为 GPU_WGPU ∧ (WIN32 ∨ X11)；
  X11Surface 增 native_display() 供 Xlib surface 双句柄同源创建
- WgpuRhi：X11 路 native_window 按 uint64_t XID 全宽传递（修 u32 截断）
- WSL 实测：ctest 299/299 全绿 + aurora_verify_x11_wgpu 自动段 18 项
  ALL PASS（XGetImage 截图物证）；Windows Release+GPU_WGPU 299/299 保持
- 修正：wgpu_x11_surface.h 头自包含（platform.h 先于门控）、win32_ime.cpp
  补平台守卫、itest_wgpu_present settle 泵消化 map/expose 同步重绘突发
- 测试设施：isolation.cpp 仓库根定位补 cwd 回退与 AURORA_REPO_ROOT 锚点
- 文档回写：规格 §8.5/§8.8、架构 §8.1/§8.4、BUILD_OPTIONS、GUIDELINE §13.2、
  README、AGENTS、CHANGELOG alpha.8

+++ src/aurora/render/gpu/wgpu_rhi.cpp
@@ -849,7 +849,8 @@ struct WgpuRhi::Impl {
         WGPUSurfaceSourceXlibWindow src{};
         src.chain.sType = WGPUSType_SurfaceSourceXlibWindow;
         src.display = options.native_display;
-        src.window = static_cast<std::uint32_t>(reinterpret_cast<std::uintptr_t>(options.native_window));
+        // webgpu.h 的 SurfaceSourceXlibWindow::window 是 uint64_t（XID 全宽），勿窄化为 u32。
+        src.window = static_cast<std::uint64_t>(reinterpret_cast<std::uintptr_t>(options.native_window));
         desc.nextInChain = &src.chain;
 #else
         AURORA_LOG_ERROR("gpu-wgpu", "native window surface not supported on this platform yet");
